Convert remainder in lib to TypeScript
